Made it a bit nicer.
Updated the names to be a little bit more descriptive,
changed public fields to private Serialized fields and added a property to change the key without having to open Visual Studio.
